Upgrade creation of idea (Hashtags in topics). Started set shares' selling price but wasn't able to finish. Going AAC vs Porto, but will be back to finish this. FORÇA AAC ALE!!!!!!!

+++ webapp/IdeaBroker/src/main/resources/model/RMI/RMI_Interface.java
@@ -35,7 +35,7 @@ public interface RMI_Interface extends Remote {
     public void updateUserTime(int uid) throws RemoteException;
     public boolean addFile(int iid, NetworkingFile file) throws RemoteException;
     public NetworkingFile getFile(int iid) throws RemoteException;
-    public void setSharesIdea(int uid, int iid, int nshares, int price)throws RemoteException;
+    public void setSharesIdea(int uid, int iid, int nshares, float price)throws RemoteException;
     boolean tryGetSharesIdea(int uid, int iid, int numShares, int targetPrice) throws RemoteException;
     boolean sayTrue() throws RemoteException;
     //public ArrayList<Notification> readNotificationsFromQueueFile(int uid) throws RemoteException;
